+++ Tests/MapboxMapsTests/Gestures/GestureManagerTests.swift
@@ -393,25 +393,77 @@ final class GestureManagerTests: XCTestCase {
     }
 
     func testOptionsPinchRotateEnabled() {
-        XCTAssertEqual(gestureManager.options.pinchRotateEnabled, true)
-        XCTAssertEqual(pinchGestureHandler.rotateEnabled, true)
+        XCTAssertTrue(gestureManager.options.pinchRotateEnabled)
+        XCTAssertTrue(pinchGestureHandler.rotateEnabled)
 
         gestureManager.options.pinchRotateEnabled = false
 
-        XCTAssertEqual(gestureManager.options.pinchRotateEnabled, false)
-        XCTAssertEqual(pinchGestureHandler.rotateEnabled, false)
+        XCTAssertFalse(gestureManager.options.pinchRotateEnabled)
+        XCTAssertFalse(pinchGestureHandler.rotateEnabled)
 
         gestureManager.options.pinchRotateEnabled = true
 
-        XCTAssertEqual(gestureManager.options.pinchRotateEnabled, true)
-        XCTAssertEqual(pinchGestureHandler.rotateEnabled, true)
+        XCTAssertTrue(gestureManager.options.pinchRotateEnabled)
+        XCTAssertTrue(pinchGestureHandler.rotateEnabled)
 
         pinchGestureHandler.rotateEnabled = false
 
-        XCTAssertEqual(gestureManager.options.pinchRotateEnabled, pinchGestureHandler.rotateEnabled)
+        XCTAssertFalse(gestureManager.options.pinchRotateEnabled)
+        XCTAssertFalse(pinchGestureHandler.rotateEnabled)
 
         pinchGestureHandler.rotateEnabled = true
 
-        XCTAssertEqual(gestureManager.options.pinchRotateEnabled, pinchGestureHandler.rotateEnabled)
+        XCTAssertTrue(gestureManager.options.pinchRotateEnabled)
+        XCTAssertTrue(pinchGestureHandler.rotateEnabled)
+    }
+
+    func testOptionsPinchZoomEnabled() {
+        XCTAssertTrue(gestureManager.options.pinchZoomEnabled)
+        XCTAssertTrue(pinchGestureHandler.zoomEnabled)
+
+        gestureManager.options.pinchZoomEnabled = false
+
+        XCTAssertFalse(gestureManager.options.pinchZoomEnabled)
+        XCTAssertFalse(pinchGestureHandler.zoomEnabled)
+
+        gestureManager.options.pinchZoomEnabled = true
+
+        XCTAssertTrue(gestureManager.options.pinchZoomEnabled)
+        XCTAssertTrue(pinchGestureHandler.zoomEnabled)
+
+        pinchGestureHandler.zoomEnabled = false
+
+        XCTAssertFalse(gestureManager.options.pinchZoomEnabled)
+        XCTAssertFalse(pinchGestureHandler.zoomEnabled)
+
+        pinchGestureHandler.zoomEnabled = true
+
+        XCTAssertTrue(gestureManager.options.pinchZoomEnabled)
+        XCTAssertTrue(pinchGestureHandler.zoomEnabled)
+    }
+
+    func testOptionsPinchPanEnabled() {
+        XCTAssertTrue(gestureManager.options.pinchPanEnabled)
+        XCTAssertTrue(pinchGestureHandler.panEnabled)
+
+        gestureManager.options.pinchPanEnabled = false
+
+        XCTAssertFalse(gestureManager.options.pinchPanEnabled)
+        XCTAssertFalse(pinchGestureHandler.panEnabled)
+
+        gestureManager.options.pinchPanEnabled = true
+
+        XCTAssertTrue(gestureManager.options.pinchPanEnabled)
+        XCTAssertTrue(pinchGestureHandler.panEnabled)
+
+        pinchGestureHandler.panEnabled = false
+
+        XCTAssertFalse(gestureManager.options.pinchPanEnabled)
+        XCTAssertFalse(pinchGestureHandler.panEnabled)
+
+        pinchGestureHandler.panEnabled = true
+
+        XCTAssertTrue(gestureManager.options.pinchPanEnabled)
+        XCTAssertTrue(pinchGestureHandler.panEnabled)
     }
 }